Better free display detection

+++ test/e2e/lib/hooks/video-recorder.js
@@ -44,6 +44,8 @@ export async function saveVideoRecording() {
 }
 
 export async function stopVideoRecording() {
+	if ( ! ffVideo ) return;
+
 	ffVideo.kill();
 	try {
 		await unlink( file );
